sigma v3 migration complete
