--- vismop-vue/src/core/custom-nodes/dashed-edge-program.ts
@@ -1,282 +1,0 @@
-/**
- * Adapted from sigma source
- * TODO: LICENSE
- */
-
-/**
- * Sigma.js WebGL Renderer Edge Program
- * =====================================
- *
- * Program rendering edges as thick lines using four points translated
- * orthogonally from the source & target's centers by half thickness.
- *
- * Rendering two triangles by using only four points is made possible through
- * the use of indices.
- *
- * This method should be faster than the 6 points / 2 triangles approach and
- * should handle thickness better than with gl.LINES.
- *
- * This version of the shader balances geometry computation evenly between
- * the CPU & GPU (normals are computed on the CPU side).
- * @module
- */
-import { floatColor, canUse32BitsIndices } from 'sigma/utils';
-import { NodeDisplayData } from 'sigma/types';
-import { ColorfadeEdgeDisplayData } from './types';
-import vertexShaderSource from './dashed-edge-vertex-shader.glsl?raw';
-import fragmentShaderSource from './dashed-edge-fragment-shader.glsl?raw';
-import { AbstractEdgeProgram } from 'sigma/rendering/webgl/programs/common/edge';
-import { RenderParams } from 'sigma/rendering/webgl/programs/common/program';
-
-const POINTS = 4;
-const ATTRIBUTES = 5;
-const STRIDE = POINTS * ATTRIBUTES;
-
-export default class DashedEdgeProgram extends AbstractEdgeProgram {
-  IndicesArray: Uint32ArrayConstructor | Uint16ArrayConstructor;
-  indicesArray: Uint32Array | Uint16Array;
-  indicesBuffer: WebGLBuffer;
-  indicesType: GLenum;
-  canUse32BitsIndices: boolean;
-  positionLocation: GLint;
-  colorLocation: GLint;
-  normalLocation: GLint;
-  matrixLocation: WebGLUniformLocation;
-  sqrtZoomRatioLocation: WebGLUniformLocation;
-  correctionRatioLocation: WebGLUniformLocation;
-  dashLengthLocation: WebGLUniformLocation;
-  gapLengthLocation: WebGLUniformLocation;
-  viewportResolutionLocation: WebGLUniformLocation;
-
-  constructor(gl: WebGLRenderingContext) {
-    super(gl, vertexShaderSource, fragmentShaderSource, POINTS, ATTRIBUTES);
-
-    // Initializing indices buffer
-    const indicesBuffer = gl.createBuffer();
-    if (indicesBuffer === null)
-      throw new Error('EdgeProgram: error while creating indicesBuffer');
-    this.indicesBuffer = indicesBuffer;
-
-    // Locations
-    this.positionLocation = gl.getAttribLocation(this.program, 'a_position');
-    this.colorLocation = gl.getAttribLocation(this.program, 'a_color');
-    this.normalLocation = gl.getAttribLocation(this.program, 'a_normal');
-
-    const matrixLocation = gl.getUniformLocation(this.program, 'u_matrix');
-    if (matrixLocation === null)
-      throw new Error('EdgeProgram: error while getting matrixLocation');
-    this.matrixLocation = matrixLocation;
-
-    const correctionRatioLocation = gl.getUniformLocation(
-      this.program,
-      'u_correctionRatio'
-    );
-    if (correctionRatioLocation === null)
-      throw new Error(
-        'EdgeProgram: error while getting correctionRatioLocation'
-      );
-    this.correctionRatioLocation = correctionRatioLocation;
-
-    const sqrtZoomRatioLocation = gl.getUniformLocation(
-      this.program,
-      'u_sqrtZoomRatio'
-    );
-    if (sqrtZoomRatioLocation === null)
-      throw new Error('EdgeProgram: error while getting sqrtZoomRatioLocation');
-    this.sqrtZoomRatioLocation = sqrtZoomRatioLocation;
-
-    const dashLengthLocation = gl.getUniformLocation(
-      this.program,
-      'u_dashLength'
-    );
-    if (dashLengthLocation === null)
-      throw new Error('EdgeProgram: error while getting dashLengthLocation');
-    this.dashLengthLocation = dashLengthLocation;
-
-    const gapLengthLocation = gl.getUniformLocation(
-      this.program,
-      'u_gapLength'
-    );
-    if (gapLengthLocation === null)
-      throw new Error('EdgeProgram: error while getting gapLengthLocation');
-    this.gapLengthLocation = gapLengthLocation;
-
-    const viewportResolutionLocation = gl.getUniformLocation(
-      this.program,
-      'u_viewportResolution'
-    );
-    if (viewportResolutionLocation === null) {
-      throw new Error(
-        'EdgeProgram: error while getting viewportResolutionLocation'
-      );
-    }
-    this.viewportResolutionLocation = viewportResolutionLocation;
-
-    // Enabling the OES_element_index_uint extension
-    // NOTE: on older GPUs, this means that really large graphs won't
-    // have all their edges rendered. But it seems that the
-    // `OES_element_index_uint` is quite everywhere so we'll handle
-    // the potential issue if it really arises.
-    // NOTE: when using webgl2, the extension is enabled by default
-    this.canUse32BitsIndices = canUse32BitsIndices(gl);
-    this.IndicesArray = this.canUse32BitsIndices ? Uint32Array : Uint16Array;
-    this.indicesArray = new this.IndicesArray();
-    this.indicesType = this.canUse32BitsIndices
-      ? gl.UNSIGNED_INT
-      : gl.UNSIGNED_SHORT;
-
-    this.bind();
-  }
-
-  bind(): void {
-    const gl = this.gl;
-
-    gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, this.indicesBuffer);
-
-    // Bindings
-    gl.enableVertexAttribArray(this.positionLocation);
-    gl.enableVertexAttribArray(this.normalLocation);
-    gl.enableVertexAttribArray(this.colorLocation);
-
-    gl.vertexAttribPointer(
-      this.positionLocation,
-      2,
-      gl.FLOAT,
-      false,
-      ATTRIBUTES * Float32Array.BYTES_PER_ELEMENT,
-      0
-    );
-    gl.vertexAttribPointer(
-      this.normalLocation,
-      2,
-      gl.FLOAT,
-      false,
-      ATTRIBUTES * Float32Array.BYTES_PER_ELEMENT,
-      8
-    );
-    gl.vertexAttribPointer(
-      this.colorLocation,
-      4,
-      gl.UNSIGNED_BYTE,
-      true,
-      ATTRIBUTES * Float32Array.BYTES_PER_ELEMENT,
-      16
-    );
-  }
-
-  computeIndices(): void {
-    const l = this.array.length / ATTRIBUTES;
-    const size = l + l / 2;
-    const indices = new this.IndicesArray(size);
-
-    for (let i = 0, c = 0; i < l; i += 4) {
-      indices[c++] = i;
-      indices[c++] = i + 1;
-      indices[c++] = i + 2;
-      indices[c++] = i + 2;
-      indices[c++] = i + 1;
-      indices[c++] = i + 3;
-    }
-
-    this.indicesArray = indices;
-  }
-
-  bufferData(): void {
-    super.bufferData();
-
-    // Indices data
-    const gl = this.gl;
-    gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, this.indicesArray, gl.STATIC_DRAW);
-  }
-
-  process(
-    sourceData: NodeDisplayData,
-    targetData: NodeDisplayData,
-    data: ColorfadeEdgeDisplayData,
-    hidden: boolean,
-    offset: number
-  ): void {
-    if (hidden) {
-      for (let i = offset * STRIDE, l = i + STRIDE; i < l; i++)
-        this.array[i] = 0;
-      return;
-    }
-
-    const thickness = data.size || 1;
-    const x1 = sourceData.x;
-    const y1 = sourceData.y;
-    const x2 = targetData.x;
-    const y2 = targetData.y;
-    const color = floatColor(data.color);
-
-    // Computing normals
-    const dx = x2 - x1;
-    const dy = y2 - y1;
-
-    let len = dx * dx + dy * dy;
-    let n1 = 0;
-    let n2 = 0;
-
-    if (len) {
-      len = 1 / Math.sqrt(len);
-
-      n1 = -dy * len * thickness;
-      n2 = dx * len * thickness;
-    }
-
-    let i = POINTS * ATTRIBUTES * offset;
-
-    const array = this.array;
-
-    // First point
-    array[i++] = x1;
-    array[i++] = y1;
-    array[i++] = n1;
-    array[i++] = n2;
-    array[i++] = color;
-
-    // First point flipped
-    array[i++] = x1;
-    array[i++] = y1;
-    array[i++] = -n1;
-    array[i++] = -n2;
-    array[i++] = color;
-
-    // Second point
-    array[i++] = x2;
-    array[i++] = y2;
-    array[i++] = n1;
-    array[i++] = n2;
-    array[i++] = color;
-
-    // Second point flipped
-    array[i++] = x2;
-    array[i++] = y2;
-    array[i++] = -n1;
-    array[i++] = -n2;
-    array[i] = color;
-  }
-
-  render(params: RenderParams): void {
-    if (this.hasNothingToRender()) return;
-
-    const gl = this.gl;
-    const program = this.program;
-
-    gl.useProgram(program);
-
-    gl.uniformMatrix3fv(this.matrixLocation, false, params.matrix);
-    gl.uniform1f(this.sqrtZoomRatioLocation, Math.sqrt(params.ratio));
-    gl.uniform1f(this.correctionRatioLocation, params.correctionRatio);
-    gl.uniform1f(this.dashLengthLocation, 3 * (1 / params.ratio));
-    gl.uniform1f(this.gapLengthLocation, 6 * (1 / params.ratio));
-    gl.uniform2f(this.viewportResolutionLocation, params.width, params.height);
-    // Drawing:
-    gl.drawElements(
-      gl.TRIANGLES,
-      this.indicesArray.length,
-      this.indicesType,
-      0
-    );
-  }
-}
